create a model hook on results route using those same parameters from the URL to create a GET request
The getJSON method runs asynchronously and returns what's called a
promise. A promise is a standard way of dealing with the results of
asynchronous operations. When getJSON receives a response from the
server, it invokes the callback function passed to .then(), and the
JSON from the response is assigned to the parameter, which we've called
responseJSON. Now, in the callback function, we can add the code that
should run after the response is received.